+++ src/main/java/com/magvy/experis/javalava_backend/application/security/config/PasswordEncoderConfig.java
@@ -1,4 +1,4 @@
-package com.magvy.experis.javalava_backend.security.config;
+package com.magvy.experis.javalava_backend.application.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
