+++ src/main/java/com/magvy/experis/javalava_backend/application/security/filter/JwtFilter.java
@@ -1,6 +1,6 @@
-package com.magvy.experis.javalava_backend.security.filter;
+package com.magvy.experis.javalava_backend.application.security.filter;
 
-import com.magvy.experis.javalava_backend.security.util.JwtUtil;
+import com.magvy.experis.javalava_backend.application.security.util.JwtUtil;
 import org.jspecify.annotations.NullMarked;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
