+++ src/main/java/com/magvy/experis/javalava_backend/controllers/AuthController.java
@@ -1,8 +1,8 @@
 package com.magvy.experis.javalava_backend.controllers;
 
 import com.magvy.experis.javalava_backend.application.DTOs.AuthDTO;
-import com.magvy.experis.javalava_backend.application.services.UserService;
-import com.magvy.experis.javalava_backend.util.JwtUtil;
+import com.magvy.experis.javalava_backend.domain.services.UserService;
+import com.magvy.experis.javalava_backend.application.security.util.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
