+++ src/main/java/com/magvy/experis/javalava_backend/domain/exceptions/MissingUserException.java
@@ -1,0 +1,7 @@
+package com.magvy.experis.javalava_backend.domain.exceptions;
+
+public class MissingUserException extends RuntimeException {
+    public MissingUserException(String message) {
+        super(message);
+    }
+}
